release 3.1.125 source code

+++ Services/GaussDBforopenGauss/V3/Model/ListDatabaseRolesResponse.php
@@ -40,7 +40,7 @@ class ListDatabaseRolesResponse implements ModelInterface, ArrayAccess
     */
     protected static $openAPIFormats = [
         'roles' => null,
-        'totalCount' => 'int64'
+        'totalCount' => 'int32'
     ];
 
     /**
